Move StringUtil to an "internal" package.
This will allow the class to be used by multiple packages.
