+++ core/src/main/java/com/google/instrumentation/stats/MeasurementDescriptor.java
@@ -14,7 +14,7 @@
 package com.google.instrumentation.stats;
 
 import com.google.auto.value.AutoValue;
-
+import com.google.instrumentation.internal.StringUtil;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
